(restore) restore and configure serve button on queue action table
