+++ packages/esm-service-queues-app/src/routes.json
@@ -129,5 +129,11 @@
       "component": "patientSearchWorkspace",
       "type": "service-queues"
     }
+  ],
+  "modals":[
+    {
+      "name": "serve-patient-modal",
+      "component": "servePatientModal"
+    }
   ]
 }
