--- packages/esm-service-queues-app/src/transition-queue-entry/transition-queue-entry-dialog.scss
@@ -1,12 +1,0 @@
-@use '@carbon/styles/scss/spacing';
-@use '@carbon/styles/scss/type';
-
-.modalBody {
-  padding-bottom: spacing.$spacing-03;
-  margin-bottom: 0rem;
-}
-
-.p {
-  margin-bottom: spacing.$spacing-02;
-  @include type.type-style('body-01');
-}
