+++ packages/esm-service-queues-app/translations/en.json
@@ -95,7 +95,6 @@
   "indication": "Indication",
   "invalidQueue": "Invalid Queue",
   "invalidtableConfig": "Invalid table configuration",
-  "lastClinicalVisit": "Last clinical visit",
   "lastEncounter": "Last encounter",
   "lastName": "Last name",
   "lastNameSort": "Last name (a-z)",
@@ -129,7 +128,6 @@
   "noLastEncounter": "There is no last encounter to display for this patient",
   "noLocationsAvailable": "No locations available",
   "noMedicationsFound": "No medications found",
-  "none": "None",
   "noNotesFound": "No notes found",
   "noPatientFoundMessage": "Sorry, no patient has been found",
   "noPatientsToDisplay": "No patients to display",
@@ -159,12 +157,10 @@
   "patientAttendingService": "Patient attending service",
   "patientHasActiveVisit": "The patient already has an active visit",
   "patientList": "Patient list",
-  "patientName": "Patient name",
   "patientNotInQueue": "The patient is not in the queue",
   "patientRemoved": "Patient removed",
   "patientRemovedFailed": "Error removing patient from queue",
   "patientRemovedSuccessfully": "Paient removed from queue successfully",
-  "patientRequeued": "Patient has been requeued",
   "patients": "Patients",
   "patientsCurrentlyInQueue": "Patients currently in queue",
   "personalDetails": "Personal Details",
@@ -221,7 +217,6 @@
   "removePatient": "Remove patient",
   "removePatientFromQueue": "Remove patient from queue",
   "removeQueueEntryError": "Error removing queue entry",
-  "requeue": "Requeue",
   "retainLocation": "Retain location",
   "returnDate": "Return date",
   "room": "Room",
@@ -267,7 +262,6 @@
   "status": "Status",
   "submitting": "Submitting...",
   "success": "Success",
-  "tcaDate": "Tca date",
   "temperature": "Temperature",
   "ticketNumber": "Ticket Number",
   "time": "Time",
